Add files via upload
this house rental management system project in PHP focuses mainly on dealing with tenants regarding their bookings and other information. Also, it displays all the available house data such as total members, address, city, state, photos, and furthermore details on it. The project is divided into two categories: Owner and Tenant Login. In an overview of this web app, the owner has full control of the system. Talking about the project, an owner can manage the booking details of all the houses that he/she lists on to the system. Also, the owner can book houses for his own too. In terms of booking for rental, both owners, as well as the tenant, have access to it by filling up certain forms.